Refactor: Consolidate CLI commands into a single file
This commit consolidates the CLI commands from separate files into a single `cli.py` file. This simplifies the project structure and improves maintainability.
